fix int to long

+++ BackEndPersona/src/main/java/sys/jp/persona/model/Persona.java
@@ -1,8 +1,6 @@
 package sys.jp.persona.model;
 
 
-import java.io.Serializable;
-
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -16,7 +14,7 @@ import jakarta.persistence.Table;
 public class Persona {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private int id;
+	private Long id;
 	private String nombre;
 	private String apellido;
 	private int edad;
@@ -28,10 +26,10 @@ public class Persona {
 	@JoinColumn (name="id_estado")
 	private Estado estado;
 	
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getNombre() {
